requirements.txt
torch>=2.0.0
transformers>=4.30.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
seqeval>=1.2.2
numpy>=1.24.0
gradio>=3.44.0
scikit-learn>=1.3.0
datasets>=2.14.0